# assiduidade_system_backend/openapi.yaml
openapi: 3.0.3
info:
  title: FaceClock API
  version: 1.0.0
  description: API principal do MVP de registro de ponto com reconhecimento facial.
servers:
  - url: https://api.faceclock.local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      required: [id, employee_code, full_name, role, status]
      properties:
        id:
          type: string
          format: uuid
        employee_code:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [COLABORADOR, GESTOR_RH, ADMIN_SISTEMA, AUDITOR]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, TERMINATED]
...
    EnrollRequest:
      type: object
      required: [user_id, captures]
      properties:
        user_id:
          type: string
          format: uuid
        captures:
          type: array
          minItems: 3
          items:
            type: object
            required: [image_base64]
            properties:
              image_base64:
                type: string
              angle:
                type: string
                example: front
    EnrollResponse:
      type: object
      required: [template_id, user_id, model_version, status]
      properties:
        template_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        model_version:
          type: string
        status:
          type: string
          enum: [ACTIVE]
    VerifyRequest:
      type: object
      required: [user_id, device_id, image_base64]
      properties:
        user_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        image_base64:
          type: string
        geo_lat:
          type: number
          format: double
        geo_lng:
          type: number
          format: double
    VerifyResponse:
      type: object
      required: [match, confidence_score, liveness_score, timestamp]
      properties:
        match:
          type: boolean
        user_id:
          type: string
          format: uuid
        confidence_score:
          type: number
          format: double
        liveness_score:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        reason:
          type: string
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Realiza autenticacao administrativa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/biometric/enroll:
    post:
      tags: [Biometric]
      security:
        - bearerAuth: []
      summary: Cadastra biometria facial do colaborador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollRequest'
      responses:
        '201':
          description: Biometria cadastrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollResponse'
        '400':
          description: Captura invalida ou consentimento ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Consentimento ativo obrigatorio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/biometric/verify:
    post:
      tags: [Biometric]
      security:
        - bearerAuth: []
      summary: Verifica biometria facial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Resultado da verificacao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '422':
          description: Imagem ou usuario invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...